AT_Parser: DATALEN is even worse than REQUESTLEN
When checking for buffer bounds, in my opinion it is
always safer to do sizeof(buffer) instead of checking
against a constant which _should_ be the buffer size.
For the same reason I would always use sizeof(*ptr)
instead of sizeof(Object) (Object *ptr).
It is just to simple to change the type of ptr and
forget to adjust the bounds check.